feat: add tweet.js file & getTweets function
